// initial-source/funky-export/var-www/disco.fan/src/types/prizelist.ts
export type prizelist = {
    listed_dex: string;
    id: number;
    token_name: string;
    symbol: string;
    ranking: number;
    quantity: number;
    price: number;
    real_probability: number;
    probability: number;
    fake_probability: number;
    saved_probability: number;
    earned_pts: number;
    decimals: number;
    ca: string;
    telegram: string;
    twitter: string;
    discord: string;
    flag: any;
    icon: string;
    default_image: string;
    balance: number;
    updatedAt: string;
    createdAt: string;
    tokenDetail: {
        id?: number;
        token_symbol?: String;
        fdv?: number;
        price?: number;
        market_cap?: number;
        circulating_supply?: number;
        total_supply?: number;
        scarcityScore?: number;
        volume_24h?: number;
        liquidity?: number;
        tradeVolumeRatio?: number;
        txns?: number;
        txns_24h?: number;
        updatedAt?: string;
        createdAt?: string;
    } ;
}